docs: Clarify that Context is the testing context not only the Juju context (#2123)
Small tweak to the `Context` docstring to clarify that it contains the
broad testing context, and that the Juju context is only part of that.

Consistently use `mgr` rather than `manager` (without re-opening the
debate about whether this should be `resource` or `event` etc).

Add the missing attributes to the documentation:
* `app_trusted`
* `juju_version`
* `exec_history`

`charm_spec` is still not included, but, although it's public, the type
is not, and I don't think we want to encourage its use (I think we would
make this private if we could). Similarly, the two capture settings are
also not included, and again don't seem to be meant for public use.