docs: explain the Charmhub public listing process and add a reference list of best practices (#1989)
**Must not be merged until the new review process is ready & active**.

### Public listing review process

The PR adds a new how-to guide
([Preview](https://canonical-ubuntu-documentation-library--1989.com.readthedocs.build/ops/1989/howto/request-public-listing/))
with instructions to get a charm publicly listed on Charmhub. Much of
this information is also provided to the reviewer and author when
requesting a public listing review, but this allows authors to
understand the process before they get to that point, and we can be (a
little!) more verbose when explaining the process.

In addition, the PR adds the collated best practices from the charming
ecosystem documentation to the "write charm code" how-to
([Preview](https://canonical-ubuntu-documentation-library--1989.com.readthedocs.build/ops/1989/howto/write-and-structure-charm-code/#best-practices)).
This is the final step in the work that distributed the best practice
notes across the documentation, and ties to the public listing review
because they form a large portion of the checklist.

*(Note that this PR sets the format of the collected best practices, but
isn't the place to change the actual wording - that should be done in
different Ops or Charmcraft PRs.)*

### Charm "maturity"

The PR also adds a new explanation page
([Preview](https://canonical-ubuntu-documentation-library--1989.com.readthedocs.build/ops/1989/explanation/charm-maturity/))
that contains general notes on charm maturity and some explanation of
the public listing review process. This replaces the [Juju Charm
Maturity](https://documentation.ubuntu.com/juju/3.6/reference/charm/charm-maturity/)
page, which will be removed after this PR is merged.

### Automatic updates

The PR adds an action and workflow action that opens a PR to refresh the
collated list of best practices, if there are additions, changes, or
removals. This is set to run weekly, which is more frequently than most
of the other scheduled workflows, but it seems unlikely that there will
be any changes most weeks.

The workflow & action are based on the 'update charm pins' action &
workflow. However, I've ended up simplifying down this action a lot
(removing all the parameters for example), and I wonder if the script
should be run from the workflow rather than bothering with having an
action. I'd appreciate feedback on this.

[Example PR for best practice
updates](https://github.com/tonyandrewmeyer/operator/pull/15), and [the
action
run](https://github.com/tonyandrewmeyer/operator/actions/runs/17111811892)
that produced it.

---------

Co-authored-by: Teodora Mihoc <teodora.mihoc@canonical.com>
Co-authored-by: Dave Wilding <tech@dpw.me>