test: Re-enable scenario consistency checks after disabling them (#2141)
This PR corrects one of the `testing` tests to correctly re-enable
consistency checks after turning them off. They default to on, and we
run our own tests with them on.

Before this PR, they would be turned off, as we check `if
os.getenv(...)`, which will be `True` when the variable is set to the
string `'false'`.