test: expand secrets integration and state transition tests (#2130)
In preparation for secrets refactor, add a good dose of secrets tests.

This PR adds one more action to the test charm, state transition and
integration tests to exercise it.
The action is parametrised with a sequence of ops secret API operations
that charm performs during a single hook.
That's designed to validate the "collapsing" or "coalescing" or secret
operations.

Done in preparation for refactoring secrets code in Ops: #1536