docs: update environment setup for integration tests and K8s tutorial (#2124)
A couple of our docs link to the Juju [environment setup
docs](https://documentation.ubuntu.com/juju/3.6/howto/manage-your-juju-deployment/set-up-your-juju-deployment-local-testing-and-development/):

1. - K8s tutorial > Set up your development environment
2. - How to write integration tests for a charm > Prepare your
environment

But actually the links point to the wrong place in the Juju docs.

For 1, I'm fixing the link.

For 2, I'm taking this opportunity to add bespoke setup instructions
that are focused on what you need for integration testing, taking
advantage of Concierge. (The Juju setup docs cover a lot more ground
than we really need)

**[Preview of new
instructions](https://canonical-ubuntu-documentation-library--2124.com.readthedocs.build/ops/2124/howto/write-integration-tests-for-a-charm/#prepare-your-environment)**