docs: change HACKING.md instructions for maintaining Charmcraft profiles (#2151)
This PR changes "Updating the Ops versions in the Charmcraft profiles"
in HACKING.md to use
[charmcraft-profile-tools](https://github.com/canonical/charmcraft-profile-tools)
for initialising charms and generating lockfile templates.

_Why not keep the manual process too? Doesn't charmcraft-profile-tools
make the steps less transparent?_ It does - but the manual process
wasn't quite right any more. The lockfiles for the K8s and machine
profiles aren't exactly the same, because the K8s profile has PyYAML as
an integration dependency. So we can't say "repeat the full process for
the machine profile" any more. Having tooling is more practical now.