docs: in integration tests, use consistent approach to logging and packing (#2150)
This PR updates [How to write integration tests for a
charm](https://documentation.ubuntu.com/ops/latest/howto/write-integration-tests-for-a-charm/)
and the [httpbin-demo
charm](https://github.com/canonical/operator/tree/main/examples/httpbin-demo)
to be consistent with the Charmcraft 4 profiles and our new machine
tutorial:

- **How to write integration tests for a charm**
- Use `sleep` to wait for Juju to process logs, then output the logs to
`stderr`.
- Change `conftest.py` to check the `CHARM_PATH` environment variable
instead of packing the charm.
- Update surrounding docs to clarify that packing should be separate
from running integration tests, and to explain how to specify which
`.charm` file to deploy in the tests.
- Make all tests depend on the `charm` fixture, for a better failure
experience if there's no `.charm` file available.

[Preview of updated
doc](https://canonical-ubuntu-documentation-library--2150.com.readthedocs.build/ops/2150/howto/write-integration-tests-for-a-charm/)

- **httpbin-demo charm** - Use `sleep` to wait for Juju to process logs,
then output the logs to `stderr`.

I haven't updated the integration tests in the Kubernetes charm
tutorial. We plan to switch the tutorial to use the Charmcraft 4 profile
at a later date.