feat: log the total number of deferred events (#2161)
When logging that a deferred notice is being processed, also log the
total number in the queue (as of the start of the hook). This allows
anyone examining the debug-level Juju log to see that there is a large
backlog, presumably indicating problems (see #2101 for an example).

There is a small cost here, in that we are collapsing a generator into a
tuple, although we expect it to be small, and we're already getting it
as a list from the storage.

Fixes #2101

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>