ci: update spelling (#2167)
The docs spellcheck seems to have updated underneath us, so make
adjustments so that the checks pass again.

For the most part, I've added words to the custom word list, but I've
made a few adjustments to the docs in cases where it seemed better than
adding the word.

The PR also corrects the description of Scenario's behaviour when
creating a temporary charm root: only the metadata files are written -
the `src` and `lib` folders are not also copied there.

(See also https://github.com/canonical/jubilant/pull/225)