docs: Add AI contribution note and style guideline for type annotation of return values (#2168)
Two small updates to our contribution and style guidelines:

* A statement on use of AI tools, summarising our discussion in Sweden.
Basically: leave the Copilot reviews to us, but you're welcome to submit
PRs that were (co-)written with AI, as long as you're a human taking
responsibility for it.
* An update to our style guide recording our decision on when a type
annotation for the return value of a function or method should be
provided.