fix: secret-info-get cannot be provided with both an ID and a label (#2170)
Ops currently silently ignores the label if both are provided by the
caller, so we need to preserve that behaviour. This is done in the
model, because `hookcmds` is a low-level API that exposes the raw Juju
errors when callers make mistakes (but the existing type overloads will
alert callers that both should not be provided).