docs: Add links to the Juju hook from each event class (#2176)
Link to the appropriate section of the Juju `Hook` from each event
class, so that all of the detailed information about the Juju event can
live in the Juju documentation, but be easily found from the Ops docs.

Also a drive-by fix to update a Charmcraft docs link from the
ReadTheDocs site to the documentation.ubuntu.com one.

Fixes #2047