refactor: use json.dumps to produce the YAML in relation-set and state-set (#2174)
`relation-set` and `state-set` both produce YAML of a simple `str:str`
dictionary that is then loaded by the Juju Go YAML library.

Rather than using `yaml.safe_dump` for this, use `json.dumps`, which
will produce JSON that is compatible YAML for this simple case. This
avoids a dependency on `ops._private.yaml`, as well as the YAML library
and the extra surface area of dumping YAML.

This is also much faster, although no-one will actually notice that of
course:

```
$ uvx --with=pyyaml --python=3.14 python -m timeit -n 50000 -s 'import yaml;a={"a":"1","b":"2","c":"3"}' 'yaml.dump(a, Dumper=yaml.CSafeDumper)'
50000 loops, best of 5: 22.1 usec per loop
$ uvx --python=3.14 python -m timeit -n 50000 -s 'import json;a={"a":"1","b":"2","c":"3"}' 'json.dumps(a)'
50000 loops, best of 5: 1.65 usec per loop
```

Suggested by @carlcsaposs-canonical , thanks!