ci: store charmcraft logs if smoke tests fail (#2192)
Sometimes the smoke test fails not in the interaction between the charm
and Juju, but before that when the charm is being packed. The charmcraft
output is almost always useless in this case, with the actual failure
buried in the charmcraft log file.

This PR updates the smoke workflow to upload the charmcraft log files if
the smoke tests fails, so that it's easier to investigate in such
situations.