feat: environment variable to control exception wrapping in testing (#2142)
This PR adds a new environment variable, `SCENARIO_BARE_CHARM_ERRORS`.
If this is set to `1` or `true` (case insensitive), then exceptions
raised in charm code in state-transition tests are re-raised as-is. If
unset or set to any other value, then the default behaviour matches the
current behaviour: charm errors are wrapped in `UncaughtCharmError` and
that is raised instead.

The idea would be to make `SCENARIO_BARE_CHARM_ERRORS=1` the default
behaviour in a future major release. In the meantime, the environment
variable allows charms to explicitly opt into the new behaviour now, for
a better debugging experience, or explicitly lock in the current
behaviour to avoid future breakages.

The `testing` tests have been updated to exercise both behaviours, while
the `ops` tests continue to assume the existing behaviour, and will
require updates when we eventually change the default.

`tox.ini` has been updated to always pass the scenario environment
variables along.

Resolves #1428.