docs: update CI examples to use uv and tox-uv (#2213)
This PR updates how tox is installed in the GitHub CI examples in our
docs. The latest machine and Kubernetes Charmcraft profiles expect uv,
tox, and tox-uv, so I figure it's good for our docs to show that too.

Unlike in our own CI, I've not pinned the uv action to a specific
commit. Should we mention something about this in the docs?