chore: linting fixes required to remove all the excludes for testing/* (#2200)
Removes all the entries in the ruff excludes that apply specifically to
`testing/*` and were left over from the unification of the ruff settings
for ops and Scenario.

In order to not have any warnings, various minor cleanup:

* Wrap some long lines.
* Add noqa's for classes and methods that copy docstrings.
* A few import order tidy-ups.
* Inside of list comps/gen-exps use `exe` rather than `exec` (the
shadowing causes no real issue, but this seems simpler than adding the
noqa).
* Add noqa's for a few exceptions inside of loops.
* In dataclasses, use `field` with `default_factory` rather than
instantiating an actual object.
* Add noqa's for a couple of uses of random, where security is not
relevant.
* Add `from None` when raising new exceptions in an exception handler.
* A few other minor adjustments.

Fixes #1693

---------

Co-authored-by: Tony Meyer <career@aotearoa.dev>