ci: add pytest.warns to note an expected warning (#2092)
Doing `stop_check` triggers a security event log, but in the Harness
tests we haven't fully configured logging, which means that a
`RuntimeWarning` is triggered. Use `pytest.warns` in the test to avoid
this polluting the test output (and also to ensure that the warning does
occur).

In addition, there's some extra cleanup of the logging. The most
significant of these is in Scenario - when the manager finishes, it
removes the JujuHandlers that were added when it was created. This seems
generally cleaner (as well as avoiding the test inconsistency) and means
that adding in the next test run is cleaner too.

---------

Co-authored-by: James Garner <james.garner@canonical.com>