docs: use base 24.04 for httpbin-demo charm (#2222)
This PR bumps the base of the
[httpbin-demo](https://github.com/canonical/operator/tree/main/examples/httpbin-demo)
charm from 22.04 to 24.04. (I'm classifying this as docs, as I consider
our example charms loosely part of the docs)

I've opened https://github.com/canonical/charmcraft/pull/2522 to make
the same change in the `kubernetes` and `machine` profiles. So I'll
leave bumping our other example charms to 24.04 until that has been
released.