ci: add ops-tracing as a dependency for the observability tests (#2239)
The workflow removes `ops` as a dependency (so that it can add it back
from main@HEAD), which also removes `ops[tracing]`. However,
`grafana-k8s` now relies on that dependency, so that breaks the tests.

This PR always adds `ops-tracing` as a dependency. `alertmanager-k8s`
and `prometheus-k8s` don't currently need that, but they have other
dependencies (like pedantic) that mean that packaging the charm should
still work.

We need a cleaner system than this, but I think #1843 may end up solving
this, and this will mean that the tests on main can pass again, so it
seems a reasonable workaround.

[Run in my
fork](https://github.com/tonyandrewmeyer/operator/actions/runs/20250813274)
- note that the failure is because of a different problem, and the
grafana-k8s tests passing is the important piece.