docs: fix URL issues by updating doc starter pack (#2238)
The "canonical" link URLs are broken in the Ops docs. For example:

```html
<link rel="canonical" href="https://documentation.ubuntu.com/ops/howto/">
```

in the page source of
https://documentation.ubuntu.com/ops/latest/howto/. The link URL is
missing `latest`, so points to a 404 page.

This PR updates to version 1.3.1 of the [doc starter
pack](https://github.com/canonical/sphinx-docs-starter-pack), which
should fix the issue.

**[Preview docs
build](https://canonical-ubuntu-documentation-library--2238.com.readthedocs.build/ops/2238/)**,
although we'll have to fully verify after merge.

**Relevant change**

In `conf.py`, set the site's base URL according to the Read the Docs
environment, instead of hard-coding.

**Other notable changes**

- Update the way that sitemap URLs are generated, to match the change in
base URLs.
- Update doc dependencies after `canonical-sphinx` was broken into
several separate Python packages (which are now more robustly developed
and tested).