ci: add scheduled workflow for packing and integration testing example charms (#2233)
As James suggested in #2229, this PR adds a scheduled workflow that
packs and runs integration tests for our example charms. I've scheduled
it to run each Tuesday before NZ hours.

(We already have a workflow that lints and runs unit tests for the
example charms. That workflow is fast and runs on PRs.)

I've organised the example charms into three jobs:
- The machine tutorial charm, based on a Charmcraft 4 profile.
- The httpbin-demo K8s charm, based on a Charmcraft 4 profile.
- The K8s tutorial charms, not based on a Charmcraft 4 profile. We're
planning to refactor these later in the 26.04 cycle.

I tested the jobs by temporarily enabling them for this PR. All are
passing after #2232 merged.